requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
--extra-index-url https://www.piwheels.org/simple

ambient @ git+https://github.com/AmbientDataInc/ambient-python-lib.git
    # via -r requirements.in
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.1.8
    # via w1thermsensor
idna==3.10
    # via requests
python-dotenv==1.1.1
    # via -r requirements.in
requests==2.32.5
    # via ambient
rpi-gpio==0.7.1
    # via -r requirements.in
sensirion-i2c-driver==1.0.0
    # via
    #   -r requirements.in
    #   sensirion-i2c-scd
sensirion-i2c-scd==0.1.2
    # via -r requirements.in
spidev==3.7
    # via -r requirements.in
urllib3==2.5.0
    # via requests
w1thermsensor==2.3.0
    # via -r requirements.in
